+++ backend/flow/src/Flow/CallbackPayload.hs
@@ -4,20 +4,20 @@ module Flow.CallbackPayload
     , RejectedEvent(..)
     , AuthenticationAttemptedResult(..)
     , AuthenticationAttemptedEvent(..)
-    , AuthenticationProviderData(..)
-    , OnfidoProviderData(..)
+    , AuthenticationProvider(..)
+    , AuthenticationProviderOnfido(..)
     )
   where
 
 import Data.Aeson
 import Data.Aeson.Casing
+import Data.Aeson.Encoding
 import Data.Aeson.Types
 import Data.Text
 import Data.Time
 import GHC.Generics
 
 import Flow.Core.Type.Callback
-import Flow.EID.AuthConfig
 import Flow.Id
 import Flow.Names
 
@@ -40,17 +40,25 @@ instance ToJSON FlowCallbackEventV1Envelope where
       eventPairs :: FlowCallbackEventV1 -> Series
       eventPairs Completed = "type" .= ("completed" :: Text)
       eventPairs Failed    = "type" .= ("failed" :: Text)
-      eventPairs (Rejected RejectedEvent { userName, mRejectMessage, mDocumentName }) =
+      eventPairs (Rejected RejectedEvent {..}) =
         ("type" .= ("flow_rejected" :: Text))
           <> ("user_name" .= userName)
-          <> ("message" .= mRejectMessage)
-          <> ("document_name" .= mDocumentName)
+          <> maybe mempty ("message" .=)       rejectMessage
+          <> maybe mempty ("document_name" .=) documentName
       eventPairs (AuthenticationAttempted AuthenticationAttemptedEvent {..}) =
-        ("type" .= ("authentication_attempted" :: Text))
-          <> ("user_name" .= userName)
-          <> ("result" .= result)
-          <> ("provider" .= provider)
-          <> ("provider_data" .= providerData)
+        "type"
+          .= ("authentication_attempted" :: Text)
+          <> "user_name"
+          .= userName
+          <> "result"
+          .= result
+          <> encodeProvider provider
+
+      encodeProvider :: AuthenticationProvider -> Series
+      encodeProvider (Onfido data') =
+        "provider" .= ("onfido" :: Text) <> ("provider_data" `pair` toEncoding data')
+      encodeProvider SmsPin = "provider" .= ("sms_pin" :: Text)
+
 
 instance FromJSON FlowCallbackEventV1Envelope where
   parseJSON = withObject "FlowCallbackEventV1Envelope" $ \o -> do
@@ -67,18 +75,25 @@ instance FromJSON FlowCallbackEventV1Envelope where
         Rejected
           <$> (   RejectedEvent
               <$> (o .: "user_name")
-              <*> (o .: "document_name")
-              <*> (o .: "message")
+              <*> (o .:? "document_name")
+              <*> (o .:? "message")
               )
-      getTypeSpecifics o "authenication_attempted" =
+      getTypeSpecifics o "authentication_attempted" =
         AuthenticationAttempted
           <$> (   AuthenticationAttemptedEvent
               <$> (o .: "user_name")
               <*> (o .: "result")
-              <*> (o .: "provider")
-              <*> (o .: "provider_data")
+              <*> getProvider o
               )
       getTypeSpecifics _ type' = fail $ "Unknown callback event type: " <> unpack type'
+
+      getProvider :: Object -> Parser AuthenticationProvider
+      getProvider o = do
+        (provider :: Text) <- o .: "provider"
+        case provider of
+          "onfido"  -> Onfido <$> o .: "provider_data"
+          "sms_pin" -> pure SmsPin
+          _         -> fail $ "Unknown AuthenticationProvider type: " <> unpack provider
 
 data FlowCallbackEventV1
     = Completed
@@ -89,44 +104,46 @@ data FlowCallbackEventV1
 
 data RejectedEvent = RejectedEvent
   { userName :: UserName
-  , mDocumentName :: Maybe DocumentName
-  , mRejectMessage ::  Maybe Text
+  , documentName :: Maybe DocumentName
+  , rejectMessage ::  Maybe Text
   }
   deriving (Eq, Generic, Show)
 
 data AuthenticationAttemptedEvent = AuthenticationAttemptedEvent
   { userName :: UserName
   , result :: AuthenticationAttemptedResult
-  , provider :: AuthProvider
-  , providerData :: AuthenticationProviderData
+  , provider :: AuthenticationProvider
   } deriving (Eq, Generic, Show)
 
-data AuthenticationAttemptedResult = Success | Failure deriving (Eq, Generic, Show)
+data AuthenticationAttemptedResult
+  = Success
+  | Failure
+  deriving (Eq, Generic, Show, Enum, Bounded)
 
-data AuthenticationProviderData = OnfidoProviderData_ OnfidoProviderData
+data AuthenticationProvider
+  = Onfido AuthenticationProviderOnfido
+  | SmsPin
   deriving (Eq, Generic, Show)
 
-data OnfidoProviderData = OnfidoProviderData { applicantId :: Text}
+newtype AuthenticationProviderOnfido = AuthenticationProviderOnfido
+  { applicantId :: Text
+  }
   deriving (Eq, Generic, Show)
+
+instance ToJSON AuthenticationProviderOnfido where
+  toEncoding = genericToEncoding aesonOptions
+
+instance FromJSON AuthenticationProviderOnfido where
+  parseJSON = genericParseJSON aesonOptions
 
 aesonOptions :: Options
-aesonOptions =
-  defaultOptions { fieldLabelModifier = snakeCase, constructorTagModifier = snakeCase }
+aesonOptions = defaultOptions { fieldLabelModifier     = snakeCase
+                              , constructorTagModifier = snakeCase
+                              , omitNothingFields      = True
+                              }
 
 instance ToJSON AuthenticationAttemptedResult where
   toEncoding = genericToEncoding aesonOptions
 
 instance FromJSON AuthenticationAttemptedResult where
-  parseJSON = genericParseJSON aesonOptions
-
-instance ToJSON AuthenticationProviderData where
-  toEncoding = genericToEncoding aesonOptions
-
-instance FromJSON AuthenticationProviderData where
-  parseJSON = genericParseJSON aesonOptions
-
-instance ToJSON OnfidoProviderData where
-  toEncoding = genericToEncoding aesonOptions
-
-instance FromJSON OnfidoProviderData where
   parseJSON = genericParseJSON aesonOptions