+++ backend/flow/src/Flow/Model/Types/Internal.hs
@@ -11,9 +11,9 @@ module Flow.Model.Types.Internal
     , InstanceSession(..)
     , InstanceKeyValues(..)
     , InstanceAccessToken(..)
-    , UserAuthConfig(..)
     , EventDetails(..)
-    , RejectionDetails(..)
+    , UserAuthenticationConfiguration(..)
+    , UserAuthenticationConfigurationData(..)
     )
  where
 
@@ -27,8 +27,8 @@ import Optics.TH
 import Auth.MagicHash
 import Auth.Session.SessionID
 import Doc.DocumentID
+import Flow.Core.Type.AuthenticationConfiguration
 import Flow.Core.Type.Callback
-import Flow.EID.AuthConfig
 import Flow.Id
 import Flow.Machinize
 import Flow.Message
@@ -133,7 +133,8 @@ data InstanceKeyValues = InstanceKeyValues
     } deriving (Eq, Generic, Show)
 
 aesonOptions :: Options
-aesonOptions = defaultOptions { fieldLabelModifier = snakeCase }
+aesonOptions =
+  defaultOptions { fieldLabelModifier = snakeCase, constructorTagModifier = snakeCase }
 
 instance FromJSON InstanceKeyValues where
   parseJSON = genericParseJSON aesonOptions
@@ -146,14 +147,24 @@ data InstanceAccessToken = InstanceAccessToken
     , instanceId :: InstanceId
     , userName :: UserName
     , hash :: MagicHash
-    }
+    } deriving (Eq, Generic, Show)
 
-data UserAuthConfig = UserAuthConfig
+data UserAuthenticationConfiguration = UserAuthenticationConfiguration
     { instanceId :: InstanceId
     , userName :: UserName
-    , authToView :: Maybe AuthConfig
-    , authToViewArchived :: Maybe AuthConfig
-    }
+    , configurationData :: UserAuthenticationConfigurationData
+    } deriving (Eq, Generic, Show)
+
+data UserAuthenticationConfigurationData = UserAuthenticationConfigurationData
+    { authenticationToView :: Maybe AuthenticationConfiguration
+    , authenticationToViewArchived :: Maybe AuthenticationConfiguration
+    } deriving (Eq, Generic, Show)
+
+instance FromJSON UserAuthenticationConfigurationData where
+  parseJSON = genericParseJSON aesonOptions
+
+instance ToJSON UserAuthenticationConfigurationData where
+  toEncoding = genericToEncoding aesonOptions
 
 makeFieldLabelsWith noPrefixFieldLabels ''Template
 makeFieldLabelsWith noPrefixFieldLabels ''InsertTemplate
@@ -166,4 +177,5 @@ makeFieldLabelsWith noPrefixFieldLabels ''FullInstance
 makeFieldLabelsWith noPrefixFieldLabels ''InstanceSession
 makeFieldLabelsWith noPrefixFieldLabels ''InstanceKeyValues
 makeFieldLabelsWith noPrefixFieldLabels ''InstanceAccessToken
-makeFieldLabelsWith noPrefixFieldLabels ''UserAuthConfig
+makeFieldLabelsWith noPrefixFieldLabels ''UserAuthenticationConfiguration
+makeFieldLabelsWith noPrefixFieldLabels ''UserAuthenticationConfigurationData