+++ backend/flow/src/Flow/Routes/Api.hs
@@ -7,7 +7,7 @@ module Flow.Routes.Api
     , CreateInstance(..)
     , RejectParam(..)
     , TemplateParameters(TemplateParameters)
-    , UserConfig(..)
+    , UserConfiguration(..)
     , InstanceEventAction(..)
     , InstanceEvent(..)
     , InstanceState(..)
@@ -22,6 +22,10 @@ module Flow.Routes.Api
     , Status(..)
     , InstanceApi
     , TemplateApi
+    , AuthenticationProvider(..)
+    , AuthenticationProviderOnfidoData(..)
+    , OnfidoMethod(..)
+    , AuthenticationConfiguration(..)
     , apiProxy
     )
   where
@@ -37,9 +41,9 @@ import Servant.API
 
 import Doc.DocumentID (DocumentID)
 import Doc.SignatoryLinkID (SignatoryLinkID)
+import Flow.Core.Type.AuthenticationConfiguration
 import Flow.Core.Type.Callback
 import Flow.Core.Type.Url
-import Flow.EID.AuthConfig
 import Flow.HighTongue
 import Flow.Id
 import Flow.Message
@@ -104,8 +108,8 @@ apiProxy :: Proxy FlowApi
 apiProxy = Proxy
 
 aesonOptions :: Options
-aesonOptions = defaultOptions { fieldLabelModifier = snakeCase }
-
+aesonOptions =
+  defaultOptions { fieldLabelModifier = snakeCase, constructorTagModifier = snakeCase }
 
 data CreateTemplate = CreateTemplate
     { name :: Text
@@ -159,7 +163,7 @@ instance ToJSON PatchTemplate where
   toEncoding = genericToEncoding aesonOptions
 
 
-data RejectParam = RejectParam
+newtype RejectParam = RejectParam
   { message :: Maybe Text
   }
   deriving (Eq, Generic, Show)
@@ -186,7 +190,7 @@ instance ToJSON CreateInstance where
 
 data TemplateParameters = TemplateParameters
   { documents :: Map DocumentName DocumentID
-  , users     :: Map UserName UserConfig
+  , users     :: Map UserName UserConfiguration
   , messages  :: Map MessageName Message
   } deriving (Eq, Generic, Show)
 
@@ -197,26 +201,26 @@ instance ToJSON TemplateParameters where
   toEncoding = genericToEncoding aesonOptions
 
 
-data UserConfig = UserConfig
+data UserConfiguration = UserConfiguration
     { flowUserId :: FlowUserId
-    , authToView :: Maybe AuthConfig
-    , authToViewArchived :: Maybe AuthConfig
+    , authenticationToView :: Maybe AuthenticationConfiguration
+    , authenticationToViewArchived :: Maybe AuthenticationConfiguration
     }
   deriving (Show, Eq, Generic)
 
-instance FromJSON UserConfig where
+instance FromJSON UserConfiguration where
   parseJSON = withObject "user config" $ \o ->
-    UserConfig
+    UserConfiguration
       <$> parseJSON (Object o)
       <*> (o .:? "auth_to_view")
       <*> (o .:? "auth_to_view_archived")
 
-instance ToJSON UserConfig where
+instance ToJSON UserConfiguration where
   toEncoding uc =
     pairs
       $  flowUserIdToSeries (flowUserId uc)
-      <> ("auth_to_view" .= authToView uc)
-      <> ("auth_to_view_archived" .= authToViewArchived uc)
+      <> ("auth_to_view" .= authenticationToView uc)
+      <> ("auth_to_view_archived" .= authenticationToViewArchived uc)
 
 data InstanceEventAction
     = Approve
@@ -226,10 +230,10 @@ data InstanceEventAction
   deriving (Eq, Ord, Generic, Show)
 
 instance FromJSON InstanceEventAction where
-  parseJSON = genericParseJSON defaultOptions { constructorTagModifier = snakeCase }
+  parseJSON = genericParseJSON aesonOptions
 
 instance ToJSON InstanceEventAction where
-  toEncoding = genericToEncoding defaultOptions { constructorTagModifier = snakeCase }
+  toEncoding = genericToEncoding aesonOptions
 
 
 data InstanceEvent = InstanceEvent
@@ -345,10 +349,10 @@ data DocumentState
   deriving (Eq, Generic, Ord, Show)
 
 instance FromJSON DocumentState where
-  parseJSON = genericParseJSON defaultOptions { constructorTagModifier = snakeCase }
+  parseJSON = genericParseJSON aesonOptions
 
 instance ToJSON DocumentState where
-  toEncoding = genericToEncoding defaultOptions { constructorTagModifier = snakeCase }
+  toEncoding = genericToEncoding aesonOptions
 
 
 data Status
@@ -358,10 +362,10 @@ data Status
   deriving (Eq, Generic, Ord, Show)
 
 instance FromJSON Status where
-  parseJSON = genericParseJSON defaultOptions { constructorTagModifier = snakeCase }
+  parseJSON = genericParseJSON aesonOptions
 
 instance ToJSON Status where
-  toEncoding = genericToEncoding defaultOptions { constructorTagModifier = snakeCase }
+  toEncoding = genericToEncoding aesonOptions
 
 
 data GetInstanceView = GetInstanceView
@@ -390,4 +394,4 @@ makeFieldLabelsWith noPrefixFieldLabels ''GetInstance
 makeFieldLabelsWith noPrefixFieldLabels ''GetInstanceView
 makeFieldLabelsWith noPrefixFieldLabels ''RejectParam
 makeFieldLabelsWith noPrefixFieldLabels ''TemplateParameters
-makeFieldLabelsWith noPrefixFieldLabels ''UserConfig
+makeFieldLabelsWith noPrefixFieldLabels ''UserConfiguration