+++ backend/flow/src/Flow/Tables.hs
@@ -133,24 +133,14 @@ tableFlowInstanceKeyValueStore = tblTable
 tableFlowUserAuthConfigs :: Table
 tableFlowUserAuthConfigs = tblTable
   { tblName        = "flow_user_auth_configs"
-  , tblVersion     = 1
-  , tblColumns     =
-    [ tblColumn { colName = "instance_id", colType = UuidT, colNullable = False }
-    , tblColumn { colName = "key", colType = TextT, colNullable = False }
-    , tblColumn { colName = "auth_to_view_provider", colType = TextT, colNullable = True }
-    , tblColumn { colName     = "auth_to_view_max_failures"
-                , colType     = IntegerT
-                , colNullable = True
-                }
-    , tblColumn { colName     = "auth_to_view_archived_provider"
-                , colType     = TextT
-                , colNullable = True
-                }
-    , tblColumn { colName     = "auth_to_view_archived_max_failures"
-                , colType     = IntegerT
-                , colNullable = True
-                }
-    ]
+  , tblVersion     = 2
+  , tblColumns = [ tblColumn { colName     = "instance_id"
+                             , colType     = UuidT
+                             , colNullable = False
+                             }
+                 , tblColumn { colName = "key", colType = TextT, colNullable = False }
+                 , tblColumn { colName = "data", colType = JsonbT, colNullable = False }
+                 ]
   , tblPrimaryKey  = pkOnColumns ["instance_id", "key"]
   , tblIndexes     = []
   , tblForeignKeys = [ (fkOnColumns ["instance_id", "key"]