+++ backend/flow/docs/changelog.md
@@ -1,5 +1,11 @@
 # API specification changelog
 
+## 0.10.6
+
+* Add Onfido authentication `method` to `AuthenticationOnfido` schema, can be
+  either `document` or `document_and_photo`.
+  `AuthenticationOnfido` is part of the `/templates/{template_id}/start` endpoint.
+
 ## 0.10.5
 
 * Add `message` parameter for flow instance rejection API.